+++ src/utility.rs
@@ -1,8 +1,8 @@
 //! Utility models submodule
-//! 
+//!
 //! This submodule defines utility functions used for serialization and deserialization
 
-use std::{fs::File, path::PathBuf};
+use std::{fs::File, path::{Path, PathBuf}};
 
 use serde_json::{Map, Value};
 use uuid::Uuid;
@@ -33,7 +33,7 @@ pub fn extract_json(
     file_name: &str,
     allowed_path: &Option<PathBuf>,
 ) -> Result<Value, ApicizeError> {
-    match get_absolute_file_name(file_name, allowed_path) {
+    match get_existing_absolute_file_name(file_name, allowed_path) {
         Ok(full_file_name) => match File::open(full_file_name) {
             Ok(file) => match serde_json::from_reader::<File, Value>(file) {
                 Ok(v) => Ok(v),
@@ -51,7 +51,7 @@ pub fn extract_csv(
     file_name: &str,
     allowed_path: &Option<PathBuf>,
 ) -> Result<Value, ApicizeError> {
-    match get_absolute_file_name(file_name, allowed_path) {
+    match get_existing_absolute_file_name(file_name, allowed_path) {
         Ok(full_file_name) => match File::open(full_file_name.clone()) {
             Ok(file) => {
                 let mut rdr = csv::Reader::from_reader(file);
@@ -65,18 +65,20 @@ pub fn extract_csv(
                 Ok(serde_json::Value::Array(data))
             }
             Err(err) => Err(ApicizeError::from_io(
-                err, Some(full_file_name.to_string_lossy().to_string()))),
+                err,
+                Some(full_file_name.to_string_lossy().to_string()),
+            )),
         },
         Err(err) => Err(err),
     }
 }
 
 /// Return the absolute file name, ensuring it exists and that it is form the same directory as our workbook
-pub fn get_absolute_file_name(
+pub fn get_existing_absolute_file_name(
     file_name: &str,
-    allowed_path: &Option<PathBuf>,
+    allowed_parent_directory: &Option<PathBuf>,
 ) -> Result<PathBuf, ApicizeError> {
-    match allowed_path {
+    match allowed_parent_directory {
         Some(parent_data_path) => {
             let data_path = PathBuf::from(parent_data_path).join(file_name);
             if data_path.exists() {
@@ -89,8 +91,49 @@ pub fn get_absolute_file_name(
             }
         }
         None => Err(ApicizeError::Error {
-            description: "External data files are unavailable in an unsaved workbook"
-                .to_string(),
+            description: "External data files are unavailable in an unsaved workbook".to_string(),
+        }),
+    }
+}
+
+/// Generate an absolute file name ensuring the directory exists
+pub fn build_absolute_file_name(
+    file_name: &str,
+    directory: &Path,
+) -> Result<PathBuf, ApicizeError> {
+    let result = PathBuf::from(directory).join(file_name);
+    if !directory.exists() {
+        Err(ApicizeError::FileAccess {
+            description: "Unable to create (invalid directory)".to_string(),
+            file_name: Some(result.to_string_lossy().to_string()),
+        })
+    } else {
+        Ok(result)
+    }
+}
+
+/// Ensure that file_path is an absolute path and is a child of allowed_parent_path, returning the relative path
+pub fn get_relative_file_name(
+    file_path: &Path,
+    parent_directory: &Path,
+) -> Result<String, ApicizeError> {
+    if !file_path.is_absolute() {
+        return Err(ApicizeError::Error {
+            description: format!(
+                "File path '{}' is not an absolute path",
+                file_path.display()
+            ),
+        });
+    }
+
+    match file_path.strip_prefix(parent_directory) {
+        Ok(relative_path) => Ok(relative_path.to_string_lossy().to_string()),
+        Err(_) => Err(ApicizeError::Error {
+            description: format!(
+                "File '{}' is not a child of '{}'",
+                file_path.display(),
+                parent_directory.display()
+            ),
         }),
     }
 }